Extract Context class
