+++ lib/datadog/appsec/waf.rb
@@ -1,20 +1,33 @@
-require 'ffi'
-require 'json'
+# frozen_string_literal: true
 
-require 'datadog/appsec/waf/result'
-require 'datadog/appsec/waf/handle'
+require 'datadog/appsec/waf/lib_ddwaf'
+
 require 'datadog/appsec/waf/version'
+require 'datadog/appsec/waf/handle'
+require 'datadog/appsec/waf/result'
+require 'datadog/appsec/waf/context'
 
 module Datadog
   module AppSec
+    # The main module exposed outside
     # rubocop:disable Metrics/ModuleLength
     module WAF
-      def self.version
+      RESULT_CODE = {
+        ddwaf_err_internal:         :err_internal,
+        ddwaf_err_invalid_object:   :err_invalid_object,
+        ddwaf_err_invalid_argument: :err_invalid_argument,
+        ddwaf_ok:                   :ok,
+        ddwaf_match:                :match,
+      }.freeze
+
+      module_function
+
+      def version
         LibDDWAF.ddwaf_get_version
       end
 
       # rubocop:disable Metrics/MethodLength,Metrics/CyclomaticComplexity,Metrics/PerceivedComplexity
-      def self.ruby_to_object(val, max_container_size: nil, max_container_depth: nil, max_string_length: nil, coerce: true)
+      def ruby_to_object(val, max_container_size: nil, max_container_depth: nil, max_string_length: nil, coerce: true)
         case val
         when Array
           obj = LibDDWAF::Object.new
@@ -137,12 +150,12 @@ module Datadog
 
           obj
         else
-          ruby_to_object(''.freeze)
+          ruby_to_object('')
         end
       end
       # rubocop:enable Metrics/MethodLength,Metrics/CyclomaticComplexity,Metrics/PerceivedComplexity
 
-      def self.object_to_ruby(obj)
+      def object_to_ruby(obj)
         case obj[:type]
         when :ddwaf_obj_invalid, :ddwaf_obj_null
           nil
@@ -174,7 +187,7 @@ module Datadog
         end
       end
 
-      def self.log_callback(level, func, file, line, message, len)
+      def log_callback(level, func, file, line, message, len)
         return if logger.nil?
 
         logger.debug do
@@ -188,11 +201,11 @@ module Datadog
         end
       end
 
-      def self.logger
+      def logger
         @logger
       end
 
-      def self.logger=(logger)
+      def logger=(logger)
         unless @log_callback
           log_callback = method(:log_callback)
           Datadog::AppSec::WAF::LibDDWAF.ddwaf_set_log_cb(log_callback, :ddwaf_log_trace)
@@ -204,122 +217,6 @@ module Datadog
         @logger = logger
       end
 
-      RESULT_CODE = {
-        ddwaf_err_internal:         :err_internal,
-        ddwaf_err_invalid_object:   :err_invalid_object,
-        ddwaf_err_invalid_argument: :err_invalid_argument,
-        ddwaf_ok:                   :ok,
-        ddwaf_match:                :match,
-      }
-
-      class Context
-        attr_reader :context_obj
-
-        def initialize(handle)
-          handle_obj = handle.handle_obj
-          retain(handle)
-
-          @context_obj = Datadog::AppSec::WAF::LibDDWAF.ddwaf_context_init(handle_obj)
-          if @context_obj.null?
-            fail LibDDWAF::Error, 'Could not create context'
-          end
-
-          validate!
-        end
-
-        def finalize
-          invalidate!
-
-          retained.each do |retained_obj|
-            next unless retained_obj.is_a?(Datadog::AppSec::WAF::LibDDWAF::Object)
-
-            Datadog::AppSec::WAF::LibDDWAF.ddwaf_object_free(retained_obj)
-          end
-
-          Datadog::AppSec::WAF::LibDDWAF.ddwaf_context_destroy(context_obj)
-        end
-
-        def run(persistent_data, ephemeral_data, timeout = LibDDWAF::DDWAF_RUN_TIMEOUT)
-          valid!
-
-          persistent_data_obj = Datadog::AppSec::WAF.ruby_to_object(
-            persistent_data,
-            max_container_size: LibDDWAF::DDWAF_MAX_CONTAINER_SIZE,
-            max_container_depth: LibDDWAF::DDWAF_MAX_CONTAINER_DEPTH,
-            max_string_length: LibDDWAF::DDWAF_MAX_STRING_LENGTH,
-            coerce: false
-          )
-          if persistent_data_obj.null?
-            fail LibDDWAF::Error, "Could not convert persistent data: #{persistent_data.inspect}"
-          end
-
-          # retain C objects in memory for subsequent calls to run
-          retain(persistent_data_obj)
-
-          ephemeral_data_obj = Datadog::AppSec::WAF.ruby_to_object(
-            ephemeral_data,
-            max_container_size: LibDDWAF::DDWAF_MAX_CONTAINER_SIZE,
-            max_container_depth: LibDDWAF::DDWAF_MAX_CONTAINER_DEPTH,
-            max_string_length: LibDDWAF::DDWAF_MAX_STRING_LENGTH,
-            coerce: false
-          )
-          if ephemeral_data_obj.null?
-            fail LibDDWAF::Error, "Could not convert ephemeral data: #{ephemeral_data.inspect}"
-          end
-
-          result_obj = Datadog::AppSec::WAF::LibDDWAF::Result.new
-          if result_obj.null?
-            fail LibDDWAF::Error, "Could not create result object"
-          end
-
-          code = Datadog::AppSec::WAF::LibDDWAF.ddwaf_run(@context_obj, persistent_data_obj, ephemeral_data_obj, result_obj, timeout)
-
-          result = Result.new(
-            RESULT_CODE[code],
-            Datadog::AppSec::WAF.object_to_ruby(result_obj[:events]),
-            result_obj[:total_runtime],
-            result_obj[:timeout],
-            Datadog::AppSec::WAF.object_to_ruby(result_obj[:actions]),
-            Datadog::AppSec::WAF.object_to_ruby(result_obj[:derivatives]),
-          )
-
-          [RESULT_CODE[code], result]
-        ensure
-          Datadog::AppSec::WAF::LibDDWAF.ddwaf_result_free(result_obj) if result_obj
-        end
-
-        private
-
-        def validate!
-          @valid = true
-        end
-
-        def invalidate!
-          @valid = false
-        end
-
-        def valid?
-          @valid
-        end
-
-        def valid!
-          return if valid?
-
-          fail LibDDWAF::Error, "Attempt to use an invalid instance: #{inspect}"
-        end
-
-        def retained
-          @retained ||= []
-        end
-
-        def retain(object)
-          retained << object
-        end
-
-        def release(object)
-          retained.delete(object)
-        end
-      end
     end
     # rubocop:enable Metrics/ModuleLength
   end
